feat: get decoder layer weights.

+++ xllm/core/runtime/worker_impl.cpp
@@ -86,7 +86,7 @@ namespace {
 class ScopedAtenLoadThreads {
  public:
   explicit ScopedAtenLoadThreads(int32_t target_threads)
-      : prev_threads_(static_cast<int32_t>(at::get_num_threads())) {
+      : prev_threads_(at::get_num_threads()) {
     if (target_threads > 0 && prev_threads_ != target_threads) {
       torch::set_num_threads(target_threads);
       active_ = true;
@@ -808,8 +808,7 @@ bool WorkerImpl::init_model(const std::string& model_weights_path,
   auto quant_args = model_loader->quant_args();
   torch::ScalarType dtype = util::parse_dtype(args.dtype(), device_);
 
-  const int64_t tokenizer_vocab_size =
-      static_cast<int64_t>(tokenizer->vocab_size());
+  const int64_t tokenizer_vocab_size = tokenizer->vocab_size();
   int64_t model_vocab_size = args.vocab_size();
   // use tokenizer vocab size if model vocab size is not set
   if (model_vocab_size <= 0) {
@@ -862,7 +861,7 @@ bool WorkerImpl::init_model(const std::string& model_weights_path,
 
   std::unique_ptr<ScopedAtenLoadThreads> scoped_load_threads;
   if (tp_world_size > 1) {
-    const int32_t prev_threads = static_cast<int32_t>(torch::get_num_threads());
+    const int32_t prev_threads = torch::get_num_threads();
     LOG(INFO) << "Temporarily setting ATen threads to 1 during weight loading"
               << ", tp_world_size=" << tp_world_size
               << ", prev_threads=" << prev_threads;
